// components/types.ts
export type navigationType = {
    id: number;
    title: string;
    path: string;
};

export type social_dataType = {
    link: string;
    className: string;
    icon: string;
};

export type SocailItemWrapperType = {
    route: any;
    children: any;
};
export type skills = {
    name: string;
    done: string;
}[];